Refactor data fetching and display in App.js, and add new components Header.js and EarthQuakeList.js
