+++ my-app/src/components/Header.js
@@ -1,0 +1,12 @@
+import React from 'react';
+
+
+function App() {
+  return (
+    <div className="App">
+      <h1>NZ Quakes</h1>
+    </div>
+  );
+}
+
+export default App;
